Refactor getActivityDetails to avoid high cyclomatic complexity
